lencre-ai-service/requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-dotenv
requests
pydantic
openai
langchain
langchain-openai
tavily-python
langdetect
eventlet
